services/client-service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.1
drf-spectacular>=0.27.2
psycopg2-binary>=2.9
pika>=1.3.2
python-dotenv>=1.0.1
